fluid clean: remove parallel to distributed.
