+++ python/paddle/distributed/fleet/utils/hybrid_parallel_util.py
@@ -17,7 +17,7 @@ from paddle import framework
 
 # (TODO: GhostScreaming) It will be removed later.
 from paddle.fluid import core
-from paddle.framework import (
+from paddle.distributed.parallel import (
     _split_tensors,
     build_groups,
     in_dygraph_mode,
