--- python/paddle/fluid/tests/unittests/collective_reducescatter.py
@@ -1,41 +1,0 @@
-# Copyright (c) 2018 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from test_collective_base import TestCollectiveRunnerBase, runtime_main
-
-import paddle
-import paddle.fluid as fluid
-
-paddle.enable_static()
-
-
-class TestCollectiveReduceScatter(TestCollectiveRunnerBase):
-    def __init__(self):
-        self.global_ring_id = 0
-
-    def get_model(self, main_prog, startup_program):
-        ring_id = 0
-        nranks = 2
-        with fluid.program_guard(main_prog, startup_program):
-            tindata = paddle.static.data(
-                name="tindata", shape=[-1, 10, 1000], dtype='float32'
-            )
-            tindata.desc.set_need_check_feed(False)
-            toutdata = fluid.layers.collective._c_reducescatter(tindata, nranks)
-            toutdata = fluid.layers.collective._c_sync_comm_stream(toutdata, 0)
-            return toutdata
-
-
-if __name__ == "__main__":
-    runtime_main(TestCollectiveReduceScatter, "reducescatter", 0)
