+++ tools/parallel_UT_rule.py
@@ -1660,7 +1660,6 @@ SIXTH_PARALLEL_JOB_NEW = [
     'test_collective_wait',
     'test_nn_matmul_v2_grad',
     'test_quant2_int8_resnet50_mkldnn',
-    'test_reducescatter_api',
     'test_collective_sendrecv',
     'test_collective_scatter',
     'test_gru_op',
@@ -1740,7 +1739,6 @@ CPU_PARALLEL_JOB = [
     'test_requantize_mkldnn_op',
     'test_repeated_fc_relu_fuse_pass',
     'test_registry',
-    'test_reducescatter_api',
     'test_reducescatter',
     'test_recurrent_op',
     'test_recommender_system',
